requirements.txt
requests==2.32.5
python-dotenv==0.9.9